+++ src/test/service/printService.js
@@ -2,8 +2,6 @@ process.env.NODE_ENV = "test";
 
 const envVariables = require("../../envVariables");
 const chai = require("chai");
-const nock = require("nock");
-const moment = require("moment");
 const chaiHttp = require("chai-http");
 chai.use(chaiHttp);
 const { expect } = chai;
@@ -18,11 +16,18 @@ const BASE_URL = "/program/v1";
 var rewire = require("rewire");
 const { getData } = require("../../service/print/dataImporter");
 const dataImporter = rewire("../../service/print/dataImporter.js");
-const pdf = rewire("../../service/print/pdf.js");
+const dataImporter1 = rewire(
+  "../../service/print/printDocx-1.0/dataImporter.js"
+);
+
+const docx = rewire("../../service/print/docx.js");
+const docx1 = rewire("../../service/print/printDocx-1.0/docx.js");
 
 const getQuestionForSection = dataImporter.__get__("getQuestionForSection");
 const getItemsFromItemset = dataImporter.__get__("getItemsFromItemset");
 const getQuestionFromItem = dataImporter.__get__("getQuestionFromItem");
+const getQuestionSet = dataImporter1.__get__("getQuestionSet");
+const getQuestionForSet = dataImporter1.__get__("getQuestionForSet");
 
 var cheerio = require("cheerio");
 var cheerioTableparser = require("cheerio-tableparser");
@@ -40,7 +45,7 @@ describe("Print Service", () => {
   });
 
   it("[Integration test] should respond with question data for correct ID", (done) => {
-    getQuestionForSection("do_1132132525993082881105")
+    getQuestionForSection("do_11341784380642918411536")
       .then((response) => {
         expect(response).to.not.be.undefined;
         expect(response.itemType).to.equal("UNIT");
@@ -53,7 +58,7 @@ describe("Print Service", () => {
   });
 
   it("[Integration test] should get Items from itemset for correct itemset ID", (done) => {
-    getItemsFromItemset("do_113213256596070400127")
+    getItemsFromItemset("do_1134178438144573441157")
       .then((response) => {
         expect(response).to.not.be.undefined;
         done();
@@ -63,32 +68,32 @@ describe("Print Service", () => {
       });
   });
 
-  it("[Integration test] should throw PDFDataImportError from itemset for incorrect itemset ID", (done) => {
+  it("[Integration test] should throw DocxDataImportError from itemset for incorrect itemset ID", (done) => {
     getItemsFromItemset("any")
       .then((response) => {
         expect(response).to.not.be.undefined;
       })
       .catch((e) => {
-        expect(e.name).to.equal("PDFDataImportError");
+        expect(e.name).to.equal("DocxDataImportError");
         expect(e.message).to.equal("Invalid Response for Itemset ID :: any");
         done();
       });
   });
 
-  it("[Integration test] should throw PDFDataImportError for incorrect item ID", (done) => {
+  it("[Integration test] should throw DocxDataImportError for incorrect item ID", (done) => {
     getQuestionFromItem("any")
       .then((response) => {
         done();
       })
       .catch((e) => {
-        expect(e.name).to.equal("PDFDataImportError");
+        expect(e.name).to.equal("DocxDataImportError");
         expect(e.message).to.equal("Invalid Response for Question ID :: any");
         done();
       });
   });
 
   it("[Integration test] should return Question Object for correct item ID", (done) => {
-    getQuestionFromItem("do_1132132526040596481722")
+    getQuestionFromItem("do_1134178438125649921264")
       .then((response) => {
         expect(response).to.not.be.undefined;
         done();
@@ -100,7 +105,7 @@ describe("Print Service", () => {
 
   it("[Integration test] should getData for correct Hierarchy ID", (done) => {
     dataImporter
-      .getData("do_11326731857693900818")
+      .getData("do_11341790341271552011559")
       .then((response) => {
         expect(response).to.not.be.undefined;
         expect(response).to.have.property("paperData");
@@ -117,10 +122,10 @@ describe("Print Service", () => {
       });
   });
 
-  xit("[Integration test] should return a PDF for correct Hierarchy ID", (done) => {
-    pdf.buildPDFWithCallback(
-      "do_11326731857693900818",
-      (base64PDF, error, errorMsg) => {
+  it("[Integration test] should return a docx for correct Hierarchy ID", (done) => {
+    docx.buildDOCXWithCallback(
+      "do_11341790341271552011559",
+      (base64, error, errorMsg) => {
         expect(error).to.be.false;
         expect(errorMsg).to.equal("");
         done();
@@ -129,14 +134,14 @@ describe("Print Service", () => {
   });
 
   it("[Integration test] should return a an error for incorrect Hierarchy ID", (done) => {
-    pdf.buildPDFWithCallback("any", (base64PDF, error, errorMsg) => {
+    docx.buildDOCXWithCallback("any", (base64, error, errorMsg) => {
       expect(error).to.be.true;
       expect(errorMsg).to.equal("Invalid ID");
       done();
     });
   });
 
-  it("[Integration test] should return and error for incorrect Hierarchy ID", (done) => {
+  it("[Integration test] should return and error for incorrect Item ID", (done) => {
     getQuestionFromItem("do_1132132526040596481722")
       .then((response) => {
         expect(response).to.not.be.undefined;
@@ -147,6 +152,87 @@ describe("Print Service", () => {
       });
   });
 
+  it("[Integration test] should return Question Object for correct questions set ID for docx1.0", (done) => {
+    getQuestionSet("do_113431918093377536172")
+      .then((response) => {
+        expect(response).to.not.be.undefined;
+        done();
+      })
+      .catch((e) => {
+        done(e);
+      });
+  });
+  it("[Integration test] docx1.0 should return and error for incorrect Hierarchy ID", (done) => {
+    getQuestionSet("do_11341847729268326411897")
+      .then((response) => {
+        expect(response).to.not.be.undefined;
+        done();
+      })
+      .catch((e) => {
+        expect(e.name).to.equal("DocxDataImportError");
+        expect(e.message).to.equal("Invalid Response for Itemset ID :: any");
+        done();
+      });
+  });
+
+  it("[Integration test] should return Question Object for correct question ID docx1.0", (done) => {
+    getQuestionForSet("do_113431952169918464189")
+      .then((response) => {
+        expect(response).to.not.be.undefined;
+        done();
+      })
+      .catch((e) => {
+        done(e);
+      });
+  });
+  it("[Integration test] docx1.0 should return and error for incorrect question ID", (done) => {
+    getQuestionForSet("do_11341 847729268326411897")
+      .then((response) => {
+        expect(response).to.not.be.undefined;
+        done();
+      })
+      .catch((e) => {
+        expect(e.name).to.equal("DocxDataImportError");
+        expect(e.message).to.equal("Invalid Response for Itemset ID :: any");
+        done();
+      });
+  });
+
+  it("[Integration test] docx1.0 should getData for correct Hierarchy ID", (done) => {
+    dataImporter1
+      .getQuestionSet("do_113431918093377536172")
+      .then((response) => {
+        expect(response).to.not.be.undefined;
+        expect(response).to.have.property("paperData");
+        expect(response).to.have.property("sectionData");
+        expect(response.sectionData).to.be.an("Array");
+        expect(response.sectionData[0].questions).to.be.an("Array");
+        expect(response).to.have.property("error");
+        expect(response.error).to.be.false;
+        done();
+      })
+      .catch((e) => {
+        done(e);
+      });
+  });
+  it("[Integration test] should return a docx for correct Hierarchy ID", (done) => {
+    docx1.buildDOCX_1_WithCallback(
+      "do_113431918093377536172",
+      (base64, error, errorMsg) => {
+        expect(error).to.be.false;
+        expect(errorMsg).to.equal("");
+        done();
+      }
+    );
+  });
+
+  it("[Integration test] should return a an error for incorrect Hierarchy ID", (done) => {
+    docx.buildDOCX_1_WithCallback("any", (base64, error, errorMsg) => {
+      expect(error).to.be.true;
+      expect(errorMsg).to.equal("Invalid ID");
+      done();
+    });
+  });
   it("Should parse table", (done) => {
     const table = `<p>Match the following:</p><figure class="table"><table><tbody><tr><td><strong>Column 1</strong></td><td><strong>Column 2</strong></td></tr><tr><td>1</td><td>1</td></tr></tbody></table></figure>`;
     $ = cheerio.load(table);
